+++ SerenityLib/examples/solarSystem/game/ships/shipSystems/ShipSystemSensors.cpp
@@ -19,7 +19,7 @@
 
 
 
-#include <core/engine/Engine.h>
+#include <core/engine/system/Engine.h>
 #include <glm/gtx/norm.hpp>
 
 using namespace Engine;
@@ -314,7 +314,7 @@ void ShipSystemSensors::update(const double& dt) {
             if (Keybinds::isPressedDownOnce(KeybindEnum::TargetCycleEnemy)) {
                 #pragma region Cycle enemies
                 const int collection_size = static_cast<int>(m_DetectedEnemyShips.size());
-                if (collection_size > 0 && collection_size > m_CurrentCycleEnemyIndex) {
+                if (collection_size > 0 && collection_size > static_cast<int>(m_CurrentCycleEnemyIndex)) {
                     auto* ship = m_DetectedEnemyShips[m_CurrentCycleEnemyIndex].ship;
                     if (ship) {
                         m_Ship.setTarget(ship, true);
@@ -346,7 +346,7 @@ void ShipSystemSensors::update(const double& dt) {
             if (Keybinds::isPressedDownOnce(KeybindEnum::TargetCycleFriendly)) {
                 #pragma region Cycle friendlies
                 const int collection_size = static_cast<int>(m_DetectedAlliedShips.size());
-                if (collection_size > 0 && collection_size > m_CurrentCycleFriendlyIndex) {
+                if (collection_size > 0 && collection_size > static_cast<int>(m_CurrentCycleFriendlyIndex)) {
                     auto* ship = m_DetectedAlliedShips[m_CurrentCycleFriendlyIndex].ship;
                     if (ship) {
                         m_Ship.setTarget(ship, true);
